add extra tests for code coverage

+++ chia/cmds/coin_funcs.py
@@ -221,7 +221,7 @@ async def async_split(
                 return []
             assert coins is not None
             assert amount_per_coin is not None
-            number_of_coins = uint64(coins[0].coin.amount // amount_per_coin.convert_amount(mojo_per_unit))
+            number_of_coins = int(coins[0].coin.amount // amount_per_coin.convert_amount(mojo_per_unit))
         elif amount_per_coin is None:
             coins = await wallet_client.get_coin_records_by_names([target_coin_id])
             if len(coins) == 0:
